Reuse shared monthly budget selector
